Fix [Artifacts ] redundant request on expanding the artifact
